页面修饰差不多了，还有一些功能上东西好像有些缺失，在修正

+++ resources/src/main/webapp/WEB-INF/pages/seedeppos.jsp
@@ -24,16 +24,16 @@
             if(userName!=""){
                 $("#sp").css("display","block");
             }
-            $("#left_depart a").mouseout(function () {
+            $("a").mouseout(function () {
                 $(this).toggleClass("a");
             });
-            $("#left_depart a").mouseover(function () {
+            $("a").mouseover(function () {
                 $(this).toggleClass("a");
             });
             $("#left_depart a").click(function () {
                 var name=$(this).html();
                 $("#departName").val(name);
-                $("#middle_position").css("display","block");
+                $(".addPosition").css("display","block");
                 $.ajax({
                     type:"post",
                     url:"seePosition",
@@ -43,10 +43,31 @@
                         for(var i=0;i<obj.length;i++){
                             $("#positions").html($("#positions").html()+"<a>"+obj[i].positionName+"</a>");
                         }
-                        $("#deleteForPos").html("<option>请选择职位</option>");
+                        $("#deleteForPos").html("<option value='0'>请选择职位</option>");
                         for(var i=0;i<obj.length;i++){
                             $("#deleteForPos").append("<option value='"+obj[i].id+"'>"+obj[i].positionName+"</option>");
                         }
+                        $("#positions a").click(function () {
+                            var posName=$(this).html();
+                            $("#positionName").val(posName);
+                            $.ajax({
+                                type:"post",
+                                url:"staffSeeStaffByPos",
+                                data:{"positionName":posName},
+                                success:function (obj1) {
+                                    $("#t_body").html("");
+                                    for(var i=0;i<obj1.length;i++){
+                                        var staff=obj1[i];
+                                        $("#t_body").html($("#t_body").html()+
+                                            "<tr><td>"+staff.id+"</td><td>"+staff.staffRealName+"</td><td>"+staff.sex+"</td><td>"+staff.age+"</td><td>"+staff.phone+"</td><td>"+staff.email+"</td><td>"+staff.staffState+"</td></tr>");
+                                    }
+                                    $("#t_body tr").click(function () {
+                                        var id=$(this).find("td").get(0);
+                                        location.href="goStaffCtrl?id="+$(id).html();
+                                    });
+                                }
+                            })
+                        });
                         $.ajax({
                             type:"post",
                             url:"staffSeeStaffByDep",
@@ -56,7 +77,7 @@
                                 for(var i=0;i<obj1.length;i++){
                                     var staff=obj1[i];
                                     $("#t_body").html($("#t_body").html()+
-                                        "<tr><td>"+staff.id+"</td><td>"+staff.staffRealName+"</td><td>"+staff.sex+"</td><td>"+staff.age+"</td><td>"+staff.phone+"</td><td>"+staff.email+"</td></tr>");
+                                        "<tr><td>"+staff.id+"</td><td>"+staff.staffRealName+"</td><td>"+staff.sex+"</td><td>"+staff.age+"</td><td>"+staff.phone+"</td><td>"+staff.email+"</td><td>"+staff.staffState+"</td></tr>");
                                 }
                                 $("#t_body tr").click(function () {
                                     var id=$(this).find("td").get(0);
@@ -64,6 +85,12 @@
                                 });
                             }
                         })
+                        $("a").mouseout(function () {
+                            $(this).toggleClass("a");
+                        });
+                        $("a").mouseover(function () {
+                            $(this).toggleClass("a");
+                        });
                     }
                 });
             });
@@ -72,10 +99,41 @@
                 location.href="goStaffCtrl?id="+$(id).html();
             });
             $("#delDep").click(function () {
+                var id=$("#departForDel option:selected").val();
+                if (id=="0"){
+                    alert("请先选择部门再删除");
+                    return;
+                }
                 location.href="deleteDep?id="+$("#departForDel option:selected").val();
             });
             $("#delPos").click(function () {
+                var id=$("#deleteForPos option:selected").val();
+                if (id=="0"){
+                    alert("请先选择职位再删除");
+                    return;
+                }
                 location.href="deletePos?id="+$("#deleteForPos option:selected").val();
+            });
+            $("#positions a").click(function () {
+                var name=$(this).html();
+                $("#positionName").val(name);
+                $.ajax({
+                    type:"post",
+                    url:"staffSeeStaffByPos",
+                    data:{"positionName":name},
+                    success:function (obj) {
+                        $("#t_body").html("");
+                        for(var i=0;i<obj.length;i++){
+                            var staff=obj[i];
+                            $("#t_body").html($("#t_body").html()+
+                                "<tr><td>"+staff.id+"</td><td>"+staff.staffRealName+"</td><td>"+staff.sex+"</td><td>"+staff.age+"</td><td>"+staff.phone+"</td><td>"+staff.email+"</td><td>"+staff.staffState+"</td></tr>");
+                        }
+                        $("#t_body tr").click(function () {
+                            var id=$(this).find("td").get(0);
+                            location.href="goStaffCtrl?id="+$(id).html();
+                        });
+                    }
+                })
             });
         })
     </script>
@@ -107,32 +165,40 @@
                     <input type="submit" value="添加部门">
                 </form>
                 <select name="departId" id="departForDel">
-                    <option>请选择部门</option>
+                    <option value="0">请选择部门</option>
                     <c:forEach items="${sessionScope.departs}" var="depart">
                         <option value="${depart.id}">${depart.departName}</option>
                     </c:forEach>
                 </select>
-                <input type="button" value="删除部门" id="delDep">
+                <input type="button" value="删除部门" id="delDep"><br><br>
                 <c:forEach items="${sessionScope.departs}" var="depart">
                     <a>${depart.departName}</a>
                 </c:forEach>
             </div>
             <div id="middle_position">
+                <div class="addPosition">
                 <form action="addPosition" method="post">
                     <input type="hidden" name="departName" id="departName">
                     职位名称：<input type="text" name="positionName"><br>
                     基本薪资：<input type="text" name="baseSal">
                     <input type="submit" value="添加职位">
                 </form>
+                </div>
                 <select name="positionId" id="deleteForPos">
-
+                    <option value="0">请选择职位</option>
+                    <c:forEach items="${sessionScope.positions}" var="position">
+                        <option value="${position.id}">${position.positionName}</option>
+                    </c:forEach>
                 </select>
-                <input type="button" value="删除职位" id="delPos">
+                <input type="button" value="删除职位" id="delPos"><br><br>
                 <div id="positions">
-
+                    <c:forEach items="${sessionScope.positions}" var="position">
+                        <a>${position.positionName}</a>
+                    </c:forEach>
                 </div>
             </div>
             <div id="right_staff">
+                <input type="hidden" id="positionName">
                 <table>
                     <tr>
                         <td>工号</td>
@@ -141,6 +207,7 @@
                         <td>年龄</td>
                         <td>手机号</td>
                         <td>邮箱</td>
+                        <td>状态</td>
                     </tr>
                     <tbody id="t_body">
                     <c:forEach items="${sessionScope.staffList}" var="staff">
@@ -151,6 +218,7 @@
                             <td>${staff.age}</td>
                             <td>${staff.phone}</td>
                             <td>${staff.email}</td>
+                            <td>${staff.staffState}</td>
                         </tr>
                     </c:forEach>
                     </tbody>
